Added app discovery to rebar3 format provider.
Default wildcards are now applied to all app directories,
and special care is taken so that behaviour is the same if the app
directory is the root directory.

Unfortunately, had to modify quite a lot of tests - first, to invoke app
discovery (as specified by the provider); second, test_app test changed
the working directory without re-updating the state, and that obviously
didn't work out great.